cross-canadian: Combine SDK and external-toolchain layer usage
Many-many changes. Sorry for the crude commit. But in general the
changes aren't that complicated. Mostly they are related with
external toolchain cross-canadian targets, some cleanups and
unifications (like setting PN in the recipes instead of the
external-toolchain-cross-canadian.bbclass similarly to the cross
class). Finally I've added a more detailed glibc libc-hedaers
file and split the package up on glibc-locale, glibc-mtrace,
glibc-scripts and glibc itself. Linux Libc headers are also in a
separate package now.

Signed-off-by: Sergey Semin <fancer.lancer@gmail.com>